Add logging and error tracking with GRPC
Added log tracking with GRPC to better handle errors.
Impacted files include server.go, client.go, and proto/nsfw.proto files.
The "exec.Command" function was updated to use "CombinedOutput" for better result handling.
Every time an error appears, it is captured and logged to the GRPC log service using the LogIssue function.
A PingMainNode function was also added to check the connection with the main node.
Added GRPC dependencies in go.sum and go.mod.
Also, .gitignore was updated to ignore the config file in the models folder.