"Update RPC URL and add Postage field in models"
In this commit, the RPC URL in the files daemon/cron.go, services/GetInscriptions.go, and apis/server.go has been updated to remove the redundant "/wallet/ord" path in the command execution. This change was done to align with the changes in recent RPC server updates.

Along with this, an additional "Postage" field has been added to the Inscriptions model in the models/inscriptions.go file to hold the postage information. This was necessary to accommodate the new feature requirements.